[MIRROR] [MDB IGNORE] Fixes a bad assignment of parent in spirit holding component (#4310)
Original PR: https://github.com/tgstation/tgstation/pull/72999
-----
## About The Pull Request

`parent` is a reference to the datum parent of the component, but this
was assigning it to a string for some reason?

I just removed it. It doesn't seem to be doing anything anyways

~~I was going to make parent a private variable but too many things
access it for some weird reason. Kinda wack~~

## Why It's Good For The Game

Bad code

## Changelog

:cl: Melbert
fix: Maybe fixes some issues with spirit holding, particularly relating
to it being in-exorcism-able.
/:cl:

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>