[MIRROR] [MDB IGNORE] Adds an ghost orbit popup for the wise cow (#4336)
Original PR: https://github.com/tgstation/tgstation/pull/73065
-----

## About The Pull Request

Ghosts now get a "hey look orbit here there's something cool going on"
alert when a wisdom cow spawns. This only applies to event spawns, not
any instance of a wisdom cow spawning.

This not being a thing has annoyed me for a LONG TIME and I've kept on
forgetting to do it.
## Why It's Good For The Game

There might not be a huge reason to orbit a wisdom cow, since it's not a
particularly juicy event or anything. Regardless, it gives deadchat
somewhere to congregate and appreciate the cow's extensive dialogue
choices before a player disintegrates it.
## Changelog
:cl:
qol: ghosts now get an orbit notification for wise cows!
/:cl:

---------

Co-authored-by: Rhials <Datguy33456@gmail.com>