[MIRROR] [MDB IGNORE] Patches Excessive "The" Out Of Hand Offering (#4341)
Original PR: https://github.com/tgstation/tgstation/pull/73040
-----

## About The Pull Request

Hey there,


![image](https://user-images.githubusercontent.com/34697715/215312944-df112bef-2a32-4206-9f46-a2b8c15145a3.png)

"her the hand" is rather silly, so let's trim it down a bit to be nice
and generic. This alert is custom-made for the hand item, and doesn't
really show up in any other context... so it should be fine? The icon
will always be that hand icon so we should be fine.

I wish we could have used the text macros like `\the` or something, but
that doesn't play too well with alerts for some reason? I don't have the
time to look into this so let's just do the simple fix of reducing the
words that show up (looks better without the larger footprint imo).
## Why It's Good For The Game

Less confusing language, better grammar and formatting.
## Changelog
:cl:
spellcheck: Whenever you offer your hand to someone in need, or just
wanting to pull them around- the alert that shows up to them should be a
bit more clear now with a 100% decrease in non-necessary pronouns and
articles.
/:cl:

---------

Co-authored-by: san7890 <the@san7890.com>
Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>