[MIRROR] [MDB IGNORE] Adds a second boom wire to payload bombs. (#4448)
Original PR: https://github.com/tgstation/tgstation/pull/73021
-----
## About The Pull Request

Adds a second detonation wire onto payload bombs, bringing the total
number of wires to 6.

## Why It's Good For The Game

As it stands, there are two wires capable of neutralizing the bomb in
most circumstances and one detonation wire. Unbolting a bomb from the
floor and spacing it is as effective a solution to a bomb as defusing it
until the explosion is imminent (You can tell by the lack of defusals
which are performed once the bomb can be moved). Even unbolting it with
10 seconds to spare gives the crew a great deal of control over where
the explosion occurs. This itself isn't an issue because shoving a bomb
out into space and hearing it explode off screen is pretty badass.
The issue is that with 2 wires capable of more or less defusing the bomb
and 1 which causes the explosion, you have a 2/3 chance to save the day
while blindly cutting wires. A second wire brings these odds to a
thrilling 50/50!
The proceed wire used to detonate the bomb when cut but will now disable
the examinable countdown timer, leaving only the beeping to indicate how
close the bomb is to exploding. This is reset once the defusal wire is
cut.

## Changelog

:cl:
balance: There's a second detonation wire on payload bombs now.
balance: Cutting the proceed wire no longer detonates the bomb but
instead deactivates the countdown timer on examine. This is reset once
the bomb is defused
/:cl:

---------

Co-authored-by: Capsandi <38051413+Capsandi@users.noreply.github.com>