[MIRROR] [MDB IGNORE] Fixes custom drink names / description being overridden (#4457)
Original PR: https://github.com/tgstation/tgstation/pull/73239
-----
## About The Pull Request

If the glass was renamed by a player, don't override the name /
description in `update_name` and `update_desc`. Simple enough

Unfortunate side effect is changing the name but not the desc leaves the
desc as it is, but I don't think anyone would really care or notice. It
used to be like this anyways.

## Why It's Good For The Game

Aesthetics. Fixes #72890

## Changelog

:cl: Melbert
fix: Custom named drinks (with a pen) no longer reset after their glass
volume updates
/:cl:

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>