[MIRROR] [MDB IGNORE] Adjusts exclamation marks to forgotten 'loud' emotes. (#4482)
Original PR: https://github.com/tgstation/tgstation/pull/73326
-----

## About The Pull Request

A player expressed disappointment that as an artificer their yell ended
with a period and not an exclamation mark. They said it did not express
the hell of being a simple mob.

I think these are the only ones that needed to be corrected. 

## Why It's Good For The Game

Exclamation marks show a heightened level of expression, heightened
expression allows for more in-depth roleplay, roleplay good.

## Changelog



:cl:
spellcheck: Adjusted punctuation marks for several emotes.
/:cl:

---------

Co-authored-by: carshalash <carshalash@gmail.com>