[MIRROR] [MDB IGNORE] Fixes soul-stealer component being broken (#4501)
Original PR: https://github.com/tgstation/tgstation/pull/73285
-----
## About The Pull Request

It extracted the first index soulstone from the list and tried to
proc-call it.

But it's a list of weakrefs, so that doesn't work. 

I just changed it to hard references. It didn't really need to be
weakrefs, since it owned it.

## Why It's Good For The Game

Useful items woo

## Changelog

:cl: Melbert
fix: The bastard sword can create constructs again. Just hit the shells
with it
/:cl:

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>