[MIRROR] [MDB IGNORE] Add all ruins to unit testing for CI (#4547)
Original PR: https://github.com/tgstation/tgstation/pull/73179
-----
## About The Pull Request
Normally ruins are RNG and only certain amount spawn for maps, but
obviously this is not ideal for testing circumstances. We want all the
ruins to spawn so the CI build doesn't have random failures. This has
been coded to only trigger for the CI pipeline and not local testing to
avoid slowing down development time for compile.

However there are some notable exceptions that are not accounted for:

- Modular mapping (too many different variations, would need to spawn
several maps with hundreds of different ruin configurations)
- Size can run out (as more ruins get added, the 255x255 map that is
used will eventually run out of room)

## Why It's Good For The Game
Better CI pipeline for testing.

## Changelog
:cl:
code: Added unit testing to force all ruins to spawn for CI build
/:cl:

---------

Co-authored-by: Tim <timothymtorres@gmail.com>