[MIRROR] [MDB IGNORE] Bumps SpacemanDMM to 1.7.3 (#4513)
Original PR: https://github.com/tgstation/tgstation/pull/73346
-----

## About The Pull Request

On the tin. You can read more about what this update entails at the
changelog here:
https://github.com/SpaceManiac/SpacemanDMM/releases/tag/suite-1.7.3 .
Just a lot of tools that were developed in 2021/2022/2023 that are
finally available for use in this way thanks to a release. And more
importantly:
## Why It's Good For The Game

Fixes #67166

oh my god this has been irritating me for so long it should
(theoretically) be fixed now since the code fixing this is now in the
release's source phew the button to click actually takes you to where
you need to go thank GOD
## Changelog
Nothing that particularly concerns players.

---------

Co-authored-by: san7890 <the@san7890.com>
Co-authored-by: Jolly-66 <70232195+Jolly-66@users.noreply.github.com>