[MIRROR] [MDB IGNORE] filled trashcarts spawn nearby grime when initialized instead of when opened for the first time (#4686)
Original PR: https://github.com/tgstation/tgstation/pull/73608
-----

## About The Pull Request
Filled trashcarts spawn some trash in them once they are opened for the
first time. But they also spawn grime, and the problem with that is that
grime spawns in a range around the trashcart. So opening the trashcart
for the first time would cause the grime to suddenly pop in around the
cart.

This PR changes it so that the grime is spawned when the filled
trashcart is initialized.
## Why It's Good For The Game
Having trash spawn around the trashcart the moment that you open it is
probably not intended. The grime is probably meant to be there before
the trashcart is opened for the first time.
## Changelog
:cl:
fix: filled trashcarts spawn nearby grime when initialized instead of
when opened for the first time
/:cl:

---------

Co-authored-by: kawoppi <94711066+kawoppi@users.noreply.github.com>
Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>