[MIRROR] [MDB IGNORE] Dangerous Research - The Alternate Sciences Research Center Space Ruin! (#4722)
Original PR: https://github.com/tgstation/tgstation/pull/73544
-----
## About The Pull Request
Adds the space ruin "dangerous_research.dmm" to the game. This ruin was
previously PR'd to two separate codebases ->
https://github.com/shiptest-ss13/Shiptest/pull/528
https://github.com/Skyrat-SS13/Skyrat-tg/pull/5047
It's a fun little lore-heavy ruin. With a few dangers and some
environmental storytelling, as well as a few documents detailing the
events that caused the science experiment to go horribly wrong.
This is absolutely the definitive version of this map, I spent a whole
lot of time redesigning and remapping it to be the gold standard for
space ruins.
## Why It's Good For The Game
We need ruins! This one is fun, has a fun/spooky/slightly edgy story to
tell, and I believe it would help bolster the few space ruins we have.
Map screenshot is below! I think it's best experienced without reading
the lore/seeing the map. If you're interested in experiencing the ruin
in-game, I'd advise against poking around too much!
<details>
<summary> Map screenshot! </summary>


![dangerous_research](https://user-images.githubusercontent.com/73589390/221678503-1235a97b-1325-42a8-ba47-3dfd12c54b86.png)


![dangerous_research2](https://user-images.githubusercontent.com/73589390/221679071-7612bd25-76b3-44cd-91a4-609e8c3b80af.png)


</details>

### Big thanks to Ferro for assistance in proofreading!

## Changelog
:cl: Cheshify, Ferro
add: A new ruin has been detected near the station, what secrets could
it hold?
/:cl:

---------

Co-authored-by: Cheshify <73589390+Cheshify@users.noreply.github.com>